Define BraveOmniboxViewViews in a separted files
