+++ browser/ui/BUILD.gn
@@ -137,6 +137,8 @@ source_set("ui") {
       "views/frame/brave_opaque_browser_frame_view.h",
       "views/frame/brave_window_frame_graphic.cc",
       "views/frame/brave_window_frame_graphic.h",
+      "views/omnibox/brave_omnibox_view_views.cc",
+      "views/omnibox/brave_omnibox_view_views.h",
       "views/reader_mode/brave_reader_mode_icon_view.cc",
       "views/reader_mode/brave_reader_mode_icon_view.h",
       "views/rounded_separator.cc",
